add utils folder, rename utils.py to helpers.py, move s3keyinfo to utils folder
